requirements.txt
discord.py
pillow
easyocr
torch
torchvision
torchaudio
